+++ Proyecto1DPOO/src/dataBase/Informe.java
@@ -4,13 +4,13 @@ import java.time.LocalDateTime; //YYYY-MM-DDTHH:MM:SS
 
 public class Informe {
 
-	private float totalJuegos;
-	private float totalPlatillos;
-	private float subTotalJuegos;
-	private float subTotalPlatillos;
-	private float totalImpuestos;
-	private float totalPropinas;
-	private float totalCosto;
+	private double totalJuegos;
+	private double totalPlatillos;
+	private double subTotalJuegos;
+	private double subTotalPlatillos;
+	private double totalImpuestos;
+	private double totalPropinas;
+	private double totalCosto;
 	private LocalDateTime fechaInicio;
 	private LocalDateTime fechaFin;
 }
